Trying to use Microsoft ML package for time series forecasting (getting exception due to project configs)

+++ CovidAnalysis/Extensions/StringExtension.cs
@@ -19,57 +19,57 @@ namespace CovidAnalysis.Extensions
             var yyyyMmDd = values[3].Split('-');
             var entryDate = new DateTime(year: int.Parse(yyyyMmDd[0]), month: int.Parse(yyyyMmDd[1]), day: int.Parse(yyyyMmDd[2]));
 
-            if (!double.TryParse(values[4], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalCasesOfSickness))
+            if (!float.TryParse(values[4], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalCasesOfSickness))
             {
                 totalCasesOfSickness = 0;
             }
 
-            if (!double.TryParse(values[10], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalCasesOfSicknessPerMillion))
+            if (!float.TryParse(values[10], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalCasesOfSicknessPerMillion))
             {
                 totalCasesOfSicknessPerMillion = 0;
             }
 
-            if (!double.TryParse(values[5], NumberStyles.Any, CultureInfo.InvariantCulture, out var newCasesOfSickness))
+            if (!float.TryParse(values[5], NumberStyles.Any, CultureInfo.InvariantCulture, out var newCasesOfSickness))
             {
                 newCasesOfSickness = 0;
             }
 
-            if (!double.TryParse(values[11], NumberStyles.Any, CultureInfo.InvariantCulture, out var newCasesOfSicknessPerMillion))
+            if (!float.TryParse(values[11], NumberStyles.Any, CultureInfo.InvariantCulture, out var newCasesOfSicknessPerMillion))
             {
                 newCasesOfSicknessPerMillion = 0;
             }
 
-            if (!double.TryParse(values[7], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalDeaths))
+            if (!float.TryParse(values[7], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalDeaths))
             {
                 totalDeaths = 0;
             }
 
-            if (!double.TryParse(values[13], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalDeathsPerMillion))
+            if (!float.TryParse(values[13], NumberStyles.Any, CultureInfo.InvariantCulture, out var totalDeathsPerMillion))
             {
                 totalDeathsPerMillion = 0;
             }
 
-            if (!double.TryParse(values[8], NumberStyles.Any, CultureInfo.InvariantCulture, out var newDeaths))
+            if (!float.TryParse(values[8], NumberStyles.Any, CultureInfo.InvariantCulture, out var newDeaths))
             {
                 newDeaths = 0;
             }
 
-            if (!double.TryParse(values[14], NumberStyles.Any, CultureInfo.InvariantCulture, out var newDeathsPerMillion))
+            if (!float.TryParse(values[14], NumberStyles.Any, CultureInfo.InvariantCulture, out var newDeathsPerMillion))
             {
                 newDeathsPerMillion = 0;
             }
 
-            if (!double.TryParse(values[35], NumberStyles.Any, CultureInfo.InvariantCulture, out var peopleVaccinated))
+            if (!float.TryParse(values[35], NumberStyles.Any, CultureInfo.InvariantCulture, out var peopleVaccinated))
             {
                 peopleVaccinated = 0;
             }
 
-            if (!double.TryParse(values[41], NumberStyles.Any, CultureInfo.InvariantCulture, out var peopleVaccinatedPerHundred))
+            if (!float.TryParse(values[41], NumberStyles.Any, CultureInfo.InvariantCulture, out var peopleVaccinatedPerHundred))
             {
                 peopleVaccinatedPerHundred = 0;
             }
 
-            if (!double.TryParse(values[46], NumberStyles.Any, CultureInfo.InvariantCulture, out var newPeopleVaccinatedSmoothedPerHundred))
+            if (!float.TryParse(values[46], NumberStyles.Any, CultureInfo.InvariantCulture, out var newPeopleVaccinatedSmoothedPerHundred))
             {
                 newPeopleVaccinatedSmoothedPerHundred = 0;
             }
