+++ CovidAnalysis/Models/CalculationItems/DTWCalcResult.cs
@@ -4,7 +4,7 @@ namespace CovidAnalysis.Models.CalculationItems
 {
     public class DTWCalcResult
     {
-        public double Cost { get; set; }
+        public float Cost { get; set; }
 
         public List<MatchingElementsModel> MatchingElements { get; set; }
     }
